fix a root persistence bug.
Before this commit, this codebase updates the acc after persisting it,
hence persisting the roots for n-1, rather than n. Since the height is
persisted right, after we restart our node, we would start from block n,
but with roots for n-1, which is incorrect and causes the node to reject
following proofs.

This is trivially solved by updating the accumulator **before**
persisting it.