Add a most stateful BestChain
We used to only save how deep is the best know chain, and nothing else.
Now we store the best block **hash**, the depth of this chain,
alternative chains and two pointers that are independent from each other
One points to the last block we validated, so we can assume all blocks
leading this one is valid. The second one is for rescan.