Add missing rules (#29)
* add consensus validation rules for transaction

* Finalize reorg logic and test it

* Make assumevalid work

* Invalidate blocks

This is the logic needed to invalidate a tip if we find an invalid block
or the user decides to.

* Handle coinbase maturity

Miners can only spend a coinbase transaction after a certain number of
confirmations. This commit verifies that, using the utreexo commitment
to this utxo height.

* Fix a problem with validation index

* Add soft-forks deployments

Until now, we validate with VERIFY_ALL, meaning we assume all soft-forks
are always enabled, but for mainnet and testnet, this isn't true.
This commit adds the activation height for buried soft-forks and only
enforces them after this block.