rpc: Find utxo (#102)
* Create a simple filter builder

* Index txid

This is useful for fetching utxos for their txid

* Implement query logic

* query script hashes, not scripts

* Move block filters to a dedicated module

* Add a simple kv database for cfilters

Use a kv bucket to store all filters on disk

* feature: allow jsonrpc finding utxos

This commit update the gettxout (and removes findtxout) from our json
rpc that now uses compact block filters to find any given utxo. We also
use the same filters to figure if the TXO is spent or not.