don't ban on the first violation
Banning a peer at the first time they misbehave (e.g. timeout a request)
might be overaggressive, and lead to us being depleted of utreexo peers.

This commit adds a banscore system. Every time a peer misbehaves we
increase that score. If it reaches a certain value, we ban them.