Assume utreexo + bugfixes (#181)
* feat: assumeutreexo

If pow fraud proofs aren't available in a given network, we can use
hard-coded assumeutreexo value. Right now we only have it for mainnet.

* fix: fix some bugs where we would get stuck

There's a bug on our functions to check if we have peers in a given
network, so we would get stuck in a few cases. This commit fixes it

* bug: fix problem address_man reads hard-coded seed

We had a silent bug where we couldn't read the hardcoded seeds due to a
json parsing error, now we sould be able to.

* add mainnet seeds

* tweak loggin on sync_node.rs

* improve reliability of our cfilters impl

* don't try to create connection too often

* implement an integrity check that doesn't go through every single block

* simplify the loggin logic for new blocks

* use the same value for assumeutreexo with or without run

* address_man: don't return any address if peer is connected

* forece to find new utreexo peers if we don't have any

* peer: only notify the node if they ack our version too

* {ibd, running}Node: do a better job at not getting stuck

* put placeholder assumeutreexo values