Wire: handle block filters out-of-order
Right now, if we receive filters out of order, we'll consider that a
misbehaving. But due to some weird connectivity or even how the remote
node processes things, we may get those in a random order. This commit
fixes that by holding the filters we've got in a BTree and only
processing it after all the previous filters already got processed.