watch-only-wallet: make the api entirely immutable
This commit changes the API for our watch-only wallet, wrapping it
inside a RwLock, so all methods take a &self ref, instead of &mut self.

This commit also replaces the watch-only instances in RPC and Electrum
to use Arc<AddressCache> instead of Arc<RwLock<AddressCache>> (i.e. we
have a lock inside the cache itself, so need to have a consumer-level lock)