watch-only-wallet: impl BlockConsumer for it
This change is intended to decouple the watch-only wallet from the
Electrum Server, in case the user don't want the latter. With this
change, you can subscribe the wallet directly, without the need to run
the Electrum Server main loop.

This commit **does not** subscribe the wallet, as we still can't disable
the Electrum Server, and it would be inefficient to notify both (the
Electrum Server needs this notification too, so it can notify connected
clients). In a future patch, we may choose whether to subscribe based on
the user's choice of running an Electrum Server or not.