cli: replace reqwest with jsonrpc (#240)
reqwest pulls in a lot of dependencies, while we only need a very simple
http client to make json-rpc requests. It also has ssl by default, using
they installed openssl. But openssl can't be statically linked due to
licensing, so if we want to build floresta for, e.g. musl, we can't
build the cli tool.

This commit replaces reqwest with the jsonrpc crate from rust-bitcoin,
reducing the dependencies tree and making it easier to static link floresta