build.rs: make it work without a git enviroment
If we try to build florestad outside the git tree, it will fail because
the build script assumes we are inside one. But while building the
Dockerfile and nix flake, we don't get the git files, so the build.rs
breaks in those cases. This commit adds a fallback, where if we can't
get the version using git, we use the one declared in our Cargo.toml.