Fix our dockerfile (once again) (#245)
* build.rs: make it work without a git enviroment

If we try to build florestad outside the git tree, it will fail because
the build script assumes we are inside one. But while building the
Dockerfile and nix flake, we don't get the git files, so the build.rs
breaks in those cases. This commit adds a fallback, where if we can't
get the version using git, we use the one declared in our Cargo.toml.

* docker: update the dockerfile build image

If we use rust's official image, we may have a problem of incompatible
libc when running on ubuntu latest, so we use the exact same image to
build, but drop the build tools after we get floresta built.

* ci: build the dockerfile on CI

This will make sure that we won't silently break the dockerfile in the future