Add criterion benches and restructure testdata (#259)
Added criterion as a dev-dependency in floresta-chain and a few initial benchmarks. Also moved the testdata directory to the root of floresta-chain, I think it is more organized and visible there.

Finally, in chain_state.rs I created a setup_test_chain function in the unit tests, and we avoid calling get_best_block a second time in the check_chain_integrity method (get both values directly instead of calling it two times).