Fix cli tests and florestad panic messages
- Make the electrum addresses be assigned by the OS to avoid conflicts (args with port `0`)
- Similarly get an available port for the rpc server and the client, on the cli side
- Improve panic messages

Finally we change `RpcImpl::create` in order to fix the "Cannot drop a runtime in a context where blocking is not allowed. This happens when a runtime is dropped from within an asynchronous context." error. When `start_http` panics (e.g. when addr is already in use), its own runtime is dropped and we see this error, which hides the error that originated the panic. This is fixed by using our tokio runtime instead.