Add test-features recipe to justfile
Created a `.sh` script using the `cargo-hack` tool in order to easily test all feature combinations. This is done in `--release` mode to reduce the size of the compiled artifacts.

The script is located in a new `contrib` directory (same approach followed by `rust-bitcoin` crates). This will be useful for CI as well.