Add clean-data recipe, rename data/ dir to tmp-db/
The `data` directory was used for ChainStore databases generated during testing. It has been renamed to `tmp-db` to explicitly signal it's a temporary log (we are using another `tests/data` directory for python tests but this one is non-generated and permanent).

A new shell script and just recipe have been added in order to delete all test-generated data. It prompts the user for confirmation with the found `tmp` and `tmp-db` directories.