rpc: fix a possible hangup with rpc
The rpcs that uses `NodeHandle` can cause our runtime to deadlock,
because we await on a sync oneshot channel to wait for the awnser. But
if the `UtreexoNode` task is running on the same worker, there's no way
it could ever resolve. This causes floresta to halt.

This commit fixes it by only calling these funtions inside a blocking
task, where blocking is fine.

A more long-term solution would be a true async handle.