Added and `install.sh` script which makes the installation processes easier (only for debian-like systems for now).
It installs apt-get/rust dependencies, build florestad/floresta_cli and
make a proper configuration of systemd service in a semi-customizable way.

It also add --assume-utreexo option as default one (so the UX can be
the fast as possible with IBD).

It also remove build dependencies after installation and configuration
of systemd service.

The only thing the user should do is to add some configurations to
florestad.service (if desired) and enable the systemd service.