metrics: add average block validation time
This metric returns the an average of how long does it take to validate
a single block. The average is computed in as a rolling fraction that
keeps a sum of the elements with a denomator telling how many blocks
we've checked. To get the average, simply dive the two and get a normal
arithmetic mean, in seconds.