Refactor `consensus.rs` error mapping
I have created a `tx_err` macro to succinctly build a `TransactionError`. This has the benefit of skipping the `BlockValidationErrors::` enum prefix when specifying the error, and flexibly taking a third argument for the inner variant value.

The validation methods now directly return `TransactionError` and use a lazily evaluated closure to compute the `txid` if there's an error.

The `pub mod error` line is now before `pub mod consensus` to make the macro available. Finally changed the "Invalid coinbase txid" error message to "Invalid Coinbase PrevOut".