Added an interactive procedure, with dialog package.
When in interactive mode, a main menu will allow the user either install
or uninstall floresta from system. If user choose to install, it will
show a menu that will help to add (or not add):

 * the type of network (bitcoin, sigtest, testnet, regtest);
 * wallet;
 * advanced setup
 * review box.

In wallet, the user will can add:

 * xpub, ypub, zpub or tpub;
 * descriptor;
 * specific address.

In advanced setup, the user can add:

 * a node to connect to (ip:port);
 * proxy (socks5 only);
 * a zeromq server to listen to (ip:port);
 * enable cfilters and filters-start-height;
 * assume-valid (with a blockheight to choose) and assume-utreexo;
 * ssl (check if openssl exists, install it if necessary and generate
PKCS#8  key and certificate paths and electrum ip/port);

After review, the user can confirm (or not confirm) the installation
procedure.