Added an interactive procedure, with dialog package.
The interactive procedure will happen if the `-N` option isnt provided.
If provided, the user need to specify them with command line options.

When in interactive mode, a main menu will allow the user either install
or uninstall floresta from system. If user choose to install, it will
show a menu that will help to add (or not add):

* the type of network (bitcoin, sigtest, testnet, regtest);

* a node to connect to (ip:port);

* proxy (socks5 only);

* a zeromq server to listen to (ip:port);

* enable cfilters and filters-start-height;

* assume-valid (with a blockheight to choose) and assume-utreexo;

* ssl (check if openssl exists, install it if necessary and generate
PKCS#8  key and certificate paths and electrum ip/port);

* addition of xpubs, descriptors and/or addresses to config.toml;; and

* review message box;

After review, the user can confirm (or not confirm) the installation
procedure.