Added test for getblockheader JSONRPC call (#398)
This commit add a structure to test calls of `getblockheader` in
tests/test_framework/floresta_rpc.py, a simple test in
tests/floresta_cli_getblockheader.py and a helper in pyproject.toml
as `florestacli-getblockheader-test`.

This simple test just call the `getblockheader` with the hash of
regtest's genesis block and expect, among some data, a prev_blockhash
full of zeros.