feat(wire): add v2 and v1 transport handler (#373)
The new transport module wraps up the V1 and V2 p2p protocols so that the rest of the app doesn't have to care, but this lead to a few refactors and open questions.

Shifted the use of RawNetworkMessage to NetworkMessage. RawNetworkMessage is essentially a V1 message (bit of a misnomer) whereas NetworkMessage is agnostic. Considering the transport module handles V1 or V2, the rest of the app doesn't care and can just use NetworkMessage.

The read and write enum types exposed by the transport module require some more explicit type bounds throughout the rest of the app. These aren't new requirements, they were just implicit before.

Pushed the TCP and SOCKS5 connection logic into the new transport module so that it can handle retries at that level. I believe this will make future transport extensions easier to implement.