Re-structured functional tests with uv tool (#402)
Optmize developer flow and the experience with tests tool through:

* a replacement of the current python package management;
* a more organized folder tree strucutre inside tests/; and
* changes in tests/run_tests.py, tests/prepare.sh and tests/run.sh to
comply with the already mentioned changes.

The replacement in python package management is from poetry to uv (a
rust-based python management tool); this approach have some PROS:

* built with rust;
* faster than poetry;
* cleanup some unecessary python dependencies;
* more readable pyproject.toml;
* compatible with python pip install -r approach (for old-school
pythonists).

Until now, a CON is:

* this remove some poetry macros. Still seems like a good replacement.

The folder structure organization allows you to define test suites in
separate folders (e.g. "florestad", "floresta-cli") without needing to
define each of them in pyproject.toml. Files ending with "-test.py" in
each folder will be recognized and executed.

To comply with above mentioned changes, tests/run_tests.py was modified
to inlcude an argument `--test-suite` that will search for a folder
`tests/<test-suite>`. It was also necessary to change `tests/prepare.sh`
and `tests/run.sh` to replace poetry to uv.

The README helps the developer to:

* install uv;
* configure functional test environment;
* format code;
* lint code;
* run tests.

It also updated .gitignore so it can help to avoid add .venv folder that
is created by uv tool.