Fix load the SSL private key as certificate and add documentation about generate keys.
The florestad/src/florestad.rs::create_tls_config function was loading
the SSL key as certificate and not as a key due to the the variable
`key_path` wrongly loading it as `self.config.ssl_cert_path` (and not as
`self.config.ssl_key_path`). Also, the --ssl-*-path options lack with a
docstring that indicate the correct way to generate those keys. So, this
commit:

* Add a docstring in both florestad.rs and cli.rs with the steps to
create the SSL private key and the certificate file;

* correctly load the SSL private key file as a private key;

* agnostically build paths to those files in a way that both UNIX and
Windows will accept;

* make a proper error handling/error-ing. Still exit, but in a clean
shutdown with a decent log.

Fix #409.