Fix load the SSL private key as certificate and add documentation about generate keys.
The florestad/src/florestad.rs::create_tls_config function was loading
the SSL key as certificate and not as a key due to the the variable
`key_path` wrongly loading it as `self.config.ssl_cert_path` (and not as
`self.config.ssl_key_path`). So, this commit:

* Add a docstring in both florestad.rs and cli.rs with the steps to
create the SSL private key and the certificate file;

* correctly load the SSL private key file as a private key;

* agnostically build paths to those files in a way that both UNIX and
Windows will accept;

* make a proper error handling/error-ing;

* warn the user when florestad cannot load TLS (files does not exist,
bad format, corrupted, or any reason of failure);

* fixed the "always log but may never listen" bug when not find a valid
SSL files (Davidson diff applied).

remove diff