Feat/enable backfill (#215)
* feat: backfill now works

On our first start-up, we can use the assume-utreexo or PoW fraud proofs to skip IBD. Each have it's own security trade-offs, but lets you start a node and get it up-and-running in no time. However, if you still want to check all blocks from genesis, but at the same time get your node up-and-running ASAP, floresta will start a background thread that will validate all blocks and check if the assumed value is valid or not. This will use a lot of CPU and bandwidth, but will not interfere with your running node.

* expose backfill to cli