CI: use corpus for fuzzing #415
The current CI test boostraps the fuzzer every time we run the CI
due to it not using any corpus. That reduces the usefulness of
running fuzz on CI. Instead of just using the fuzzer like this,
we should use some corpus that already have a decent coverage
and will stress our code in a satisfactory way.

This commit uses the corpus data provided by the floresta-qa-assets
repository. It contains corpus data generated over months of continuous
fuzzing.