Added --gen-selfsigned-cert flag (#425)
* Added --gen-selfsigned-cert flag

Fix #423

This flag, when passed to florestad, will create a ssl folder in
<data-dir> and <data-dir>/ssl/key.pem and a
self-signed<data-dir>/ssl/cert.pem.

This self-signed certificate can be a option to users that do not have
certificates with a trusted chain. It's important to note that the
private-key isn't encrypted.

More specifically, this commit add:

* a flag --gen-selfsigned-cert on CLI;

* some enums elements on error.rs;

* a generate_selfsigned_certificate on florestad.rs;

* a new step on florestad::start method before creating the
TLS-enabledelectrum server.

* Added entry on README about the SSL certificates

The readme give some instructions about node:

* with a non-TLS enabled electrum;
* with a TLS enabled electrum given a PKCS#8 private-key / certificate;
* with a TLS enabled electrum with self-signed PKCS#8 private-key /
certificate;