Create `UtxoData` and check coinbase maturity (#440)
This adds `UtxoData` struct from PR #300, but with an additional `is_coinbase` flag. We get the flag from the utreexo leaf header code, just like the utxo creation height. The utxo creation unix time is not computed nor used anywhere (covered at #300).

Most of this diff is refactoring `TxOut` to `UtxoData` in functions. I'm generally not using a `UtxoMap` type alias in order to keep the map types explicit on a first look.

The `Consensus::verify_transaction` now checks the spent coinbase tx outputs maturity. `ChainParams` has a constant for the maturity period but it's always 100 regardless of the network.