running_node: don't ask for blocks until utreexo peers announce it
Currently, if we are notified for a new block, we ask for headers and
then try to download the actual data. One problem may emerge if our
utreexo peers haven't seen the block (or are still processing it). We
may ask for a block that they don't have, making them take too long to
respond and get punished for it. Eventually, we will ban all our utreexo
peers, thanks to that.

This commit changes that logic to only attempt to download a new block
iff it gets announced by our utreexo peers; or if a while has passed
without anything new (in this case our utreexo peers may be withholding a block from us).