[fix #428] refactor README into doc folder (#429)
This PR addresses issue #428 by making the README.md more concise and clear, refactoring it into a doc folder.

Now we have a document for each topic, that can be expanded upon to better explain its subject. The readme is concise and only contains the most useful information about the project.