Add `testdata` utxo files with `UtxoData` (#444)
I have pre-computed the full `UtxoData` vector for the two current test blocks. This way we can test every aspect of the block, including coinbase UTXOs maturity, relative timelocks, etc. This replaces the previous `TxOut` vectors.

We also make the `UtxoData` struct de/serializable if the `test-utils` feature is set. This is used for benches to have the serde methods available.