chain: Replace bitcoinconsensus with bitcoinkernel
libbitcoinconsensus is deprecated, and have a suboptimal performance due
to deserializing transactions for every input, even if we are validating
them in sequence. Libbitcoinkernel has a wrapper type for transactions,
and we pass a & ref to it. We can then create the tx once and pass it as
reference to validate each input. This causes a massive perf gain during
script validation.

This commit replaces all mentions to libbitcoinconsensus, and replaces
with kernel where neede. Notice that we can't use both, because of name
clashing on linking time.