wire: rework how dns seeds work
The current code for dns seeds is very frustrating: it asks for
addresses even when we don't need, it takes a while to return (holding
the node meanwhile), can't be disabled.

This commit reworks that to fix those problems. The dns seed code now
only runs after we've been more than one minute and we haven't seen at
least one connection. It runs on the background and notifies the node
back one it's finished.

I've also added an option to disable the dns
seeds. This could be useful for privacy reasons (don't want to leak to
my dns provider that I use bitcoin).