[RPC] getblockchaininfo core alike
This commit reforms the rpc command getblockchaininfo to be alike cores, seeking for floresta to be a
drop-in replacement for btc core.

Also, two new commands are added;

1. getutreexostate which returns the state of the utreexo accumulator, this one have all the utreexo related camps
that getblockchaininfo had.

2. getbackfillstate which returns some info about backfill, this one have all the backfill related camps that
getblockchaininfo had.

The rest of the changes are; test for the new commands (TODO), new structures to satisfy the new information
required.