wire: fix a small bug with handle_disconnection (#474)
If a peer disconnects, we re-do all inflight requests that we had with
that peer. However, the code that does so may fail early if
`redo_inflight_request` return an error. It is the case if we request
for utreexo data or compact block filters without having a peer that
supports those requests. This commit adds an early check that makes
these functions act as a no-op in case we can't fulfill the request.

It is not a problem to not re-request stuff here, since we already have
some fallback logic that will try it again.