Small refactors to remove unwraps (#475)
A couple of replacements of `unwrap` for `expect` with proper messages.

In partial_chain.rs `get_validation_flags` now takes the hash directly from the caller, avoiding the `unwrap` (exactly as we do for `get_validation_flags` in chain_state.rs).

In `get_leaf_hashes` (udata.rs) we instead take the TxOut directly. We also avoid computing the txid twice by passing it.