flat-store: check for DB corruption
This commit adds a simple, xxHash-based check for integrity for the
flat_chain_store. If we find it to be corrupted, we reindex to find how
much of it is lost, and try to continue from there.

There's no error-correction logic in this commit, because this tends to
bring complexity and require more space. Since modern file-systems are
so good at keeping things from corrupting, I don't think there's that
much need to make our own error correction code here.