Deny unwraps in `floresta-chain`, handle errors (#478)
We deny the `clippy::unwrap_used` for non test code. Then we use expects if it's infallible and no `Result` is returned. If it's infallible but we can return it's clearer to use `?`.

`verify_script` is refactored to return `Result` with chainstore errors, but other return-database-error refactors are `TODO`.

Finally we were returning `BlockNotPresent` errors in cases where the block is actually present but it's `Orphan` or `InvalidChain`. Created a new error variant for these two cases. It's also used in places where we were using `disk_header.height().unwrap()`.