Remove unused `florestad` and `wire` features (#486)
The last usage of the `florestad/experimental-p2p` feature was removed at #209. It is not used anywhere now.

Also we have a `floresta-wire/pruned_utreexo_chainstate` feature which is not used either. I guess we could reintroduce it if in the future we have a non pruned chain (but maybe it is not needed since we work with trait bounds).