chain: fix a bug with connect_block
While running `florestad` I've found a weird bug where the `SyncNode` would never
make progress, dispite having the blocks available in the hash map. It
turns out the bug was caused by an unclean shutdown, where the
validation index AND accumulator was pointing to a block we had as
`FullyValid`. If we see a block like this, we return early and don't do
anything. But this means our validation index will never make progress.

This commit fixes this by handling the special case where even though we
already validated that block, our validation index points to it. So we
won't return early and do update the state.