wire: don't ask for blocks if we already have many
If we fail to start processing blocks for some reason, the backlog of
unprocessed blocks will grow really fast, causing the node to take up a
lot of ram, until it crashes. This commit problem is particularly bad in
the `SyncNode`.

This commit fixes it by adding a check in the block request logic,
requiring our backlog to be smaller than 10 before we can donwload more.
Notice that this doesn't relate to timed-out requests.