chain: handle accumulator reorg
Right now, if we ever encounter ourselves in a reorg of the chain, we
will get into a corrupted state, because we won't update our accumulator
for the no most-work-chain. Therefore, causing us to get stuck.

This commit adds logic to save an retrieve accumulators from different
heights, allowing a full reorg of our chain.