Optimize `FlatChainStore` and improve tests (#493)
Previously we were doing:

```rust
    unsafe fn get_header_by_hash(
        &self,
        hash: BlockHash,
    ) -> Result<Option<DiskBlockHeader>, FlatChainstoreError> {
        let header = self
            .block_index
            .get_index_for_hash(hash, |height| self.get_block_header_by_index(height))?
            .map(|height| self.get_block_header_by_index(height));

        Ok(header.transpose()?.map(|x| x.header))
    }
```

Which we can optimize/simplify by making the `get_index_for_hash` function directly return the header as well, which is free since it's already fetched.

To make this easier I make the `IndexBucket` enum contain the header if occupied, and named the fields for clarity.

In the `ChainStore::get_block_hash` impl we handle the `NotFound` case. Then I have improved the headers tests by verifying our assumptions are holding.