Bugfix: fix two bugs that may happen after an unclean shutdown (#488)
* chain: fix a bug with connect_block

While running `florestad` I've found a weird bug where the `SyncNode` would never
make progress, dispite having the blocks available in the hash map. It
turns out the bug was caused by an unclean shutdown, where the
validation index AND accumulator was pointing to a block we had as
`FullyValid`. If we see a block like this, we return early and don't do
anything. But this means our validation index will never make progress.

This commit fixes this by handling the special case where even though we
already validated that block, our validation index points to it. So we
won't return early and do update the state.

* wire: don't ask for blocks if we already have many

If we fail to start processing blocks for some reason, the backlog of
unprocessed blocks will grow really fast, causing the node to take up a
lot of ram, until it crashes. This commit problem is particularly bad in
the `SyncNode`.

This commit fixes it by adding a check in the block request logic,
requiring our backlog to be smaller than 10 before we can donwload more.
Notice that this doesn't relate to timed-out requests.