wire: prefer v2 connections for non-utreexo peers
We already have a good quorum of people running nodes that support v2,
we can prefer v2 connections over v1 at this point safely. The only
exception for now is `utreexod`, that still doesn't have v2, but we
need to connect with them do download proofs.

This commit adds a logic that turns the v1 fallback off for all
non-utreexo peers.