wire: prefer v2 connections for non-utreexo peers
We already have a good quorum of people running nodes that support v2,
we can prefer v2 in most cases without any problem. The only problem is
that bridge nodes (mainly utreexod) doesn't support BIP0324 yet, meaning
that if we toggle the `no-v1-fallback` option on, we'll never download
any block. For that reason, this commit adds an exception for `utreexo`
nodes, where they can use V1, even if `no-v1-fallback` is set.

For RPC, if the `v2transport` cli option is set, we disallow V1
fallback.