Make ChainStore methods take `&mut self` (#500)
The rationale for this is that we shouldn't assume implementations will use interior mutability. It can also be dangerous if the type auto-implements `Sync`, as two threads could try to mutate and read the same data concurrently.

I created an `update_header_and_index` helper function to properly manage the write lock.

Also improved slightly the `test_reorg` to make the tested scenario more clear to readers.