wire: prefer v2 connections for non-utreexo peers (#501)
* wire: prefer v2 connections for non-utreexo peers

We already have a good quorum of people running nodes that support v2,
we can prefer v2 in most cases without any problem. The only problem is
that bridge nodes (mainly utreexod) doesn't support BIP0324 yet, meaning
that if we toggle the `no-v1-fallback` option on, we'll never download
any block. For that reason, this commit adds an exception for `utreexo`
nodes, where they can use V1, even if `no-v1-fallback` is set.

For RPC, if the `v2transport` cli option is set, we disallow V1
fallback.

* florestad: make `allow-v1-fallback` false by default

This commit inverts the `no-v1-fallback` to `allow-v1-fallback`. The new
option continues to be default `false`, but now it means that we won't
option V1 connections (expect with utreexod) by default.

* cli: document the behavior of v2transport

This commit adds a short description of what the `v2transport` cli
argument means

* functional-tests: fix the addnode test

This commit swaps Utreexod for Bitcoin Core in this test. For the V1
tests we start core with the `v2transport` option as false, so it won't
accept v2 transactions.

By doing so, we can test both the behavior with and without v2 support.