Use `&mut self` in FlatChainStore inner methods (#512)
The methods that conceptually mutate the state (even if using raw pointers) should take `&mut self` to let the compiler enforce the ownership rules correctly. This way it will be harder for us to misuse the methods. We also get rid of the `#[allow(clippy::mut_from_ref)]`.

The only exception is `BlockIndex.set_index_for_hash` (because we pass a `&self` closure to it). This should be fine as we only use it inside `add_index_entry`.

Polished a bit the docstrings and fixed two conditional compilation warnings for Windows.