wire: update the block proof message flow
Until now, utreexo proofs were sent along with a block. When
downloading a block, you would request for a utreexo block, which,
has the actual block + some utreexo data. While this works (and
is simpler to implement), it doesn't allow for proof caching on the
requester-side.

For that reason, we've changed the message flow to separate blocks
and proofs. A client will request a block, look at which UTXOs were
spent there, filter for what they have in cache, and then request
proof and leaf data they don't have.

This commit implements this in floresta. It won't, however, cache
stuff. Rather, we ask for all the inputs that weren't created on
the same block. A UTXO cache is a TODO.