chain: handle accumulator reorg (#498)
* chain: handle accumulator reorg

Right now, if we ever encounter ourselves in a reorg of the chain, we
will get into a corrupted state, because we won't update our accumulator
for the no most-work-chain. Therefore, causing us to get stuck.

This commit adds logic to save an retrieve accumulators from different
heights, allowing a full reorg of our chain.

* func-tests: add a test for reorg

This commit adds a functional test for reorg, where we:
 - Create an utreexod and florestad instance
 - Connect them
 - Mine 10 blocks using utreexod
 - Wait for florestad to catch up
 - Use `invalidateblock` on utreexod to cause a fork on block 5
 - Mine 20 more blocks on top of the tip
 - Check if both have the same tip and accumulator at the end