florestad: make FlatChainstore the default
This commit changes the default chain store to FlatChainstore, rather
than KvChainstore.

To avoid also compiling the unused KvChainstore every time, I've hidden
it behind a kv-chainstore feature. For tests, or in florestad, you
shouldn't enable both at the same time, or this will cause a compilation
error. This is due to the compilation-time selection of chainstore
causing naming clashing if both are defined. For that reason, I had to
change the feature_matrix.sh file to treat flat-chainstore and
kv-chainstore as mutually excluside but required (i.e. you always have to
define exactly one of them).