Improve leaf reconstruction error handling
Previously we were returning a simple `Error` enum with only one variant (`EmptyStack`) as leaf reconstruction error. This variant was used for cases where the script was actually not empty.

I added the remaining 2 variants, and then a higher level error type that we use at `process_proof` to provide context.

If we ever get any of these leaf errors, we will now see the error data debugged (the compact leaf we were trying to reconstruct, the txin, and the error kind).