Fix just lint recipe (#538)
This recipe was previously checking no features and all features, which should cover most cases and is cheap (while we fully enforce clippy on the more expensive `just lint-features` and on CI).

But now the `kv-chainstore` and `flat-chainstore` features are mutually exclusive, so this recipe is broken. I have updated it to properly handle the particular cases of `floresta-chain` and `florestad`. Since cargo doesn't have something like `--all-features-except`, I had to list all the features and exclude one chainstore feature.