tests: fix the addnode func test
I've changed the approach to check if the peer is down, by sending
one ping, wait a little, then try sending another ping. The first
one won't succeed, and an attempt to wire a new one will cause the
socket to err. This actually has a smaller timeout than the
original, so the test runs faster than before.