Lint `fuzz` and `metrics`
I added these two packages to the crate list in `feature_matrix` so that we lint them (and test them if in the future we add tests).

The `just lint` recipe already checks them since it is run with the `--workspace` flag (and `just fmt` uses the `--all` flag). This change is to enforce clippy at CI.

I updated the benchmarking docs, and also fixed an issue with the `cargo-fuzz` installation.