Wire: small fixes and improvements (#537)
* request_blocks: don't ask for blocks that we already have

* sync_node: change the default times to avoid overloading slower peers

* peer: rename Timeout to PingTimeout

* peer: log if we get a UtreexoBlock

* sync_node: ask for less blocks in parallel, as this may overload our peers

* sync_node: ask for blocks we've missed for some reason

* wire: don't force the last block requested on disconnection

We will re-do this request in `handle_disconnection`, this could create duplicate requests

* transport: use BufReader to improve read performance, specially with slow connections

* tests: fix the addnode func test

I've changed the approach to check if the peer is down, by sending
one ping, wait a little, then try sending another ping. The first
one won't succeed, and an attempt to wire a new one will cause the
socket to err. This actually has a smaller timeout than the
original, so the test runs faster than before.

* wire: get rid of magic values

This commit removes the magic values we use while deciding when and
how many blocks to download. I've created two constants on
 "NodeContext" that modulates this. One is how many blocks we ask
on a single GETDATA message, and one for how many of those messages
we can have in parallel.

* sync-node: remove duplicate field

We already have a last_block_request on `UtreexoNode`, no need to
have another one in `SyncNode`.