Benchmark checksums and chainstore creation (#553)
We have found that the checksum method reads all the virtual data. It may be useful to benchmark it, which helps evaluating future changes. Also the initialization time for each chainstore.

`DbCheckSum` and the method will now be public, although it seems harmless as the inner fields are private (i.e., the only way to produce a `DbCheckSum` is via the method).