floresta-chain: fix a test feature problem (#557)
If we ran cargo test without any features, it turns both
chainstores on, because `floresta-electrum` has `kv-chainstore`
as dev dep, but `florestad` has `flat-chainstore`.

This in turn, causes trouble with our tests, since we have
tests with both chinstore, just overloading the setup functions
at compilation time. Therefore, we get a name clash.

It isn't a problem for the test matrix, because we told cargo-hack
that `flat` and `kv` are mutually exclusive.