running_node: call maybe_create_connection
We are using `create_connection` directly, but this function assumes
you've checked whether we are not connected with our `--connec` peer.
If we don't check it, we will see the `PeerAlreadyExists` error over
and over again. `maybe_create_connection` also helps using the
fixed peers and dns seeds if we don't have any available addresses
to work with

Therefore, I'm replacing `create_connection` with `maybe_create_connection`
for `check_connections`.